Fwd: [PATCH v6.6-RT] kvm: Use the proper mask to check for the LAZY bit.
kvm: Use the proper mask to check for the LAZY bit.

To check for the LAZY bit in ti_work, it is needed to use
_TIF_NEED_RESCHED_LAZY which contains the shifted bit (the mask) instead
of TIF_NEED_RESCHED_LAZY which contains the bit number.

This broke while switch from the old PREEMPT_LAZY code to PREEMPT_AUTO.

Use _TIF_NEED_RESCHED_LAZY instead of TIF_NEED_RESCHED_LAZY.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Clark Williams <williams@redhat.com>